requirements.txt
beautifulsoup4==4.6.0
praw==5.3.0
python-dotenv==0.7.1
